Update preset handling for Pools/Miners
If new presets for a pool or a miner arrive in the code, new sections will be added to the user-config files miners.config.txt and pools.config.txt automatically.

+++ Include.psm1
@@ -1684,25 +1684,32 @@ function Set-MinersConfigDefault {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory = $True)]
-        [String]$PathToFile
+        [String]$PathToFile,
+        [Parameter(Mandatory = $False)]
+        [Switch]$Force = $false
     )
-    try {
-        $Done = [PSCustomObject]@{}
-        $Devices = Select-Device @(Get-Device "gpu") -Type @("nvidia","amd") | Select-Object Model,Vendor -Unique | Foreach-Object {$_ | Add-Member Vendor $(Get-DeviceVendor $_) -Force;$_}
-        $Setup = Get-ChildItemContent ".\Data\MinersConfigDefault.ps1" | Select-Object -ExpandProperty Content
-        $Setup.PSObject.Properties | Where-Object Membertype -eq NoteProperty | Select-Object Name,Value | Foreach-Object {
-            $Setup_Name = $_.Name
-            $Setup_Content = [PSCustomObject[]]$_.Value
-            $VendorSet = $false
-            $Devices | Foreach-Object {
-                if (-not $VendorSet) {$Done | Add-Member "$($Setup_Name)-$($_.Vendor)" $Setup_Content;$VendorSet=$true}
-                $Done | Add-Member "$($Setup_Name)-$($_.Model)" $Setup_Content        
+    if ($Force -or -not (Test-Path $PathToFile) -or (Get-ChildItem $PathToFile).LastWriteTime.ToUniversalTime() -lt (Get-ChildItem ".\Data\MinersConfigDefault.ps1").LastWriteTime.ToUniversalTime()) {
+        try {
+            if (Test-Path $PathToFile) {$Preset = Get-Content $PathToFile | ConvertFrom-Json}
+            $Done = [PSCustomObject]@{}
+            $Devices = Select-Device @(Get-Device "gpu") -Type @("nvidia","amd") | Select-Object Model,Vendor -Unique | Foreach-Object {$_ | Add-Member Vendor $(Get-DeviceVendor $_) -Force;$_}
+            $Setup = Get-ChildItemContent ".\Data\MinersConfigDefault.ps1" | Select-Object -ExpandProperty Content
+            $Setup.PSObject.Properties | Where-Object Membertype -eq NoteProperty | Select-Object Name,Value | Foreach-Object {
+                $Setup_Name = $_.Name
+                $Setup_Content = [PSCustomObject[]]$_.Value
+                $VendorSet = $false
+                $Devices | Foreach-Object {
+                    if (-not $VendorSet) {"$($Setup_Name)-$($_.Vendor)";$VendorSet=$true}
+                    "$($Setup_Name)-$($_.Model)"
+                } | Foreach-Object {
+                    $Done | Add-Member $_ $(if ($Preset -and $Preset.PSObject.Properties.Name -icontains $_){$Preset.$_}else{$Setup_Content})
+                }
             }
+            $Done | ConvertTo-Json | Set-Content $PathToFile -Encoding utf8
         }
-        $Done | ConvertTo-Json | Set-Content $PathToFile -Encoding utf8
-    }
-    catch{
-        Write-Log -Level Error "Could not create $($PathToFile) "
+        catch{
+            Write-Log -Level Error "Could not create $($PathToFile) "
+        }
     }
 }
 
@@ -1710,29 +1717,38 @@ function Set-PoolsConfigDefault {
     [CmdletBinding()]
     param(
         [Parameter(Mandatory = $True)]
-        [String]$PathToFile
+        [String]$PathToFile,
+        [Parameter(Mandatory = $False)]
+        [Switch]$Force = $false
     )
-    try {
-        $Done = [PSCustomObject]@{}
-        $Setup = Get-ChildItemContent ".\Data\PoolsConfigDefault.ps1" | Select-Object -ExpandProperty Content
-        Get-ChildItem ".\Pools\*.ps1" | Select-Object -ExpandProperty BaseName | Foreach-Object {        
-            $Setup_Content = [PSCustomObject]@{}
-            $Setup_Currencies = @("BTC")
-            if ($Setup.$_) {
-                if ($Setup.$_.Fields) {$Setup_Content = $Setup.$_.Fields}
-                $Setup_Currencies = @($Setup.$_.Currencies)            
+    if ($Force -or -not (Test-Path $PathToFile) -or (Get-ChildItem $PathToFile).LastWriteTime.ToUniversalTime() -lt (Get-ChildItem ".\Data\PoolsConfigDefault.ps1").LastWriteTime.ToUniversalTime()) {
+        try {
+            if (Test-Path $PathToFile) {$Preset = Get-Content $PathToFile | ConvertFrom-Json}
+            $Done = [PSCustomObject]@{}
+            $Setup = Get-ChildItemContent ".\Data\PoolsConfigDefault.ps1" | Select-Object -ExpandProperty Content
+            Get-ChildItem ".\Pools\*.ps1" | Select-Object -ExpandProperty BaseName | Foreach-Object {        
+                if (-not $Preset -or $Preset.PSObject.Properties.Name -inotcontains $_) {
+                    $Setup_Content = [PSCustomObject]@{}                
+                    $Setup_Currencies = @("BTC")
+                    if ($Setup.$_) {
+                        if ($Setup.$_.Fields) {$Setup_Content = $Setup.$_.Fields}
+                        $Setup_Currencies = @($Setup.$_.Currencies)            
+                    }
+                    if ($Setup_Content.PSObject.Properties.Keys -inotcontains "Worker") {$Setup_Content | Add-Member Worker "`$WorkerName" -Force}
+                    if ($Setup_Content.PSObject.Properties.Keys -inotcontains "Penalty") {$Setup_Content | Add-Member Penalty 0 -Force}
+                    if ($Setup_Content.PSObject.Properties.Keys -inotcontains "Algorithm") {$Setup_Content | Add-Member Algorithm "" -Force}
+                    if ($Setup_Content.PSObject.Properties.Keys -inotcontains "ExcludeAlgorithm") {$Setup_Content | Add-Member ExcludeAlgorithm "" -Force}            
+                    $Setup_Currencies | Foreach-Object {$Setup_Content | Add-Member $_ "$(if ($_ -eq "BTC"){"`$Wallet"})" -Force}
+                } else {
+                    $Setup_Content = $Preset.$_
+                }
+                $Done | Add-Member $_ $Setup_Content
             }
-            if ($Setup_Content.PSObject.Properties.Keys -inotcontains "Worker") {$Setup_Content | Add-Member Worker "`$WorkerName" -Force}
-            if ($Setup_Content.PSObject.Properties.Keys -inotcontains "Penalty") {$Setup_Content | Add-Member Penalty 0 -Force}
-            if ($Setup_Content.PSObject.Properties.Keys -inotcontains "Algorithm") {$Setup_Content | Add-Member Algorithm "" -Force}
-            if ($Setup_Content.PSObject.Properties.Keys -inotcontains "ExcludeAlgorithm") {$Setup_Content | Add-Member ExcludeAlgorithm "" -Force}            
-            $Setup_Currencies | Foreach-Object {$Setup_Content | Add-Member $_ "$(if ($_ -eq "BTC"){"`$Wallet"})" -Force}
-            $Done | Add-Member $_ $Setup_Content
+            $Done | ConvertTo-Json | Set-Content $PathToFile -Encoding utf8
         }
-        $Done | ConvertTo-Json | Set-Content $PathToFile -Encoding utf8
-    }
-    catch{
-        Write-Log -Level Error "Could not create $($PathToFile) "
+        catch{
+            Write-Log -Level Error "Could not create $($PathToFile) "
+        }
     }
 }
